requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.8.0
fire==0.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
Naked==0.1.32
packaging==23.0
pdf2docx==0.5.1
Pillow==9.5.0
platformdirs==2.5.2
PyJWT==2.6.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==2.2.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.16.5
